+++ server/src/tests/cases/compiler/error1.c
@@ -1,0 +1,11 @@
+int test() {
+    error("This is an error message");    
+}
+
+/**
+ * this tests the error() call expr handling in the binder
+ * If successful, error should act as a return point of the function
+ * Which means it won't report an error that the function doesn't return a value
+ */
+
+// @driver: fluffos